+++ apps/common/src/main/scala/com/daml/network/config/AutomationConfig.scala
@@ -31,4 +31,7 @@ case class AutomationConfig(
     /** Only intended for testing. Allows disabling governance automation.
       */
     enableSvcGovernance: Boolean = true,
+    /** Only intended for testing. Allows disabling governance automation.
+      */
+    enableCometbftReconciliation: Boolean = true,
 ) {}