+++ apps/sv/src/main/scala/com/daml/network/sv/config/SvAppConfig.scala
@@ -151,8 +151,6 @@ case class SvAppClientConfig(
 }
 case class CometBftConfig(
     enabled: Boolean = false,
-    // TODO(#5740): Remove the option to disable automation once we're ready to pull that trigger
-    automationEnabled: Boolean = false,
     connectionUri: String = "",
 )